Add int/turbo, fix flow control, support remote
Adds int and turbo modes to the fan speed selector.

Moves the ESP to ID 12. It's not clear if all remotes use the same ID or not, we only have one datapoint, so keep an eye on this and consider waiting for the first ping to complete on startup to dedupe. For now just assume 12 is probably fine.

Fix flow control. This was a bit of an adventure, as there are several rules involved. Seems to work in the single unit test case (eg, no alignment errors once it picks up) but will need data from a setup with a remote to see how that fares.

Misc code cleanup.